Add files via upload
iTunes search interface written with python. User enters a search term and the script will search for songs, movies, and other media that matches that term. The user is then presented with a numbered preview of the results. The user can enter the number of the item they would like to preview, enter a new search, or exit. If the user selects an item, the item's page on the iTunes website is launched.